Added helmet system to game

+++ core/src/com/apricotjam/spacepanic/components/helmet/HelmetScreenComponent.java
@@ -5,12 +5,12 @@ import com.badlogic.ashley.core.Component;
 import com.badlogic.gdx.utils.ObjectMap;
 
 public class HelmetScreenComponent implements Component {
-	public ObjectMap<Resource, Integer> resourceCount = new ObjectMap<Resource, Integer>();
+	public ObjectMap<Resource, Float> resourceCount = new ObjectMap<Resource, Float>();
 	
 	public HelmetScreenComponent() {
-		resourceCount.put(Resource.OXYGEN, 25);
-		resourceCount.put(Resource.OIL, 20);
-		resourceCount.put(Resource.RESOURCE2, 15);
-		resourceCount.put(Resource.RESOURCE3, 10);
+		resourceCount.put(Resource.OXYGEN, 25.0f);
+		resourceCount.put(Resource.OIL, 20.0f);
+		resourceCount.put(Resource.RESOURCE2, 15.0f);
+		resourceCount.put(Resource.RESOURCE3, 10.0f);
 	}
 }
